# backend/hhrr_app/static/swagger/swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"description": "This API provides the Plena Global Holdings software development team with the necessary knowledge base for the integration of the different applications with the services offered by the human resources department.",
		"version": "1.0.0",
		"title": "HUMMAN RESOURCES DEPARTMENT API",
		"license": {
			"name": "MIT",
			"url": "https://opensource.org/licenses/MIT"
		}
	},
	"servers": [
		{
			"url": "/api/v1/"
		}
	],
	"components": {
		"schemas": {
...
			"tbl_demographic_data": {
				"type": "object",
				"properties": {
					"ccn_demographic_data": {
						"type": "integer",
						"example": 1,
						"description": "Incremental integer number assigned by the system automatically when recording a new record."
					},
					"ccn_employee": {
						"type": "integer",
						"minimum": 1,
						"maximum": 10000,
						"example": 1,
						"description": "Foreign key with table tbl_employee"
					},
					"birth_department": {
						"type": "number",
						"minimum": 1,
						"maximum": 10000,
						"example": 1,
						"description": "Foreign key with table tbl_department"
					},
					"birth_city": {
						"type": "number",
						"minimum": 1,
						"maximum": 10000,
						"example": 1,
						"description": "Foreign key with table tbl_city"
					},
					"department_residence": {
						"type": "number",
						"minimum": 1,
						"maximum": 10000,
						"example": 1,
						"description": "Foreign key with table tbl_department"
					},
					"city_residence": {
						"type": "number",
						"minimum": 10,
						"maximum": 10000,
						"example": 1,
						"description": "Foreign key with table tbl_city"
					},
					"ccn_schooling_level": {
						"type": "string",
						"minimum": 10,
						"maximum": 10000,
						"example": 1,
						"description": "Foreign key with table tbl_schooling_level"
					},
					"ccn_race": {
						"type": "number",
						"minimum": 1,
						"maximum": 10000,
						"example": 1,
						"description": "Foreign key with table tbl_race"
					},
					"is_head_of_household": {
						"type": "date",
						"example": "1990-04-14",
						"description": "Check list for save if the employee is head of household"
					}
				}
			},
...
			"tbl_auto_perceived_gender": {
				"type": "object",
				"properties": {
					"ccn_auto_perceived_gender": {
						"type": "integer (pk) auto_increment",
						"example": 1,
						"description": "Incremental integer number assigned by the system automatically when recording a new record."
					},
					"auto_perceived_gender": {
						"type": "number",
						"minimum": 0,
						"maximum": 40,
						"example": "Femenino",
						"description": "String Field"
					}
				}
			},
...
		}
	},
	"paths": {
...
		"/auto_perceived_gender": {
			"post": {
				"tags": [
					"Basic data from auto perceived gender"
				],
				"description": "Create a new gender in the human resources platform",
				"requestBody": {
					"description": "Post New gender to database",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/tbl_auto_perceived_gender"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "CREATED",
						"schema": {
							"$ref": "#/components/schemas/tbl_auto_perceived_gender"
						}
					},
					"404": {
						"description": "NOT FOUND",
						"schema": {
							"$ref": "#/components/schemas/tbl_auto_perceived_gender"
						}
					}
				}
			},
			"get": {
				"tags": [
					"Basic data from auto perceived gender"
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/components/schemas/tbl_auto_perceived_gender"
						}
					},
					"404": {
						"description": "NOT FOUND",
						"schema": {
							"$ref": "#/components/schemas/tbl_auto_perceived_gender"
						}
					}
				}
			}
		},
		"/auto_perceived_gender/{ccn_auto_perceived_gender}": {
			"get": {
				"tags": [
					"Basic data from auto perceived gender"
				],
				"parameters": [
					{
						"name": "ccn_auto_perceived_gender",
						"in": "path",
						"description": "todo id to update",
						"required": true,
						"type": "integer"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/components/schemas/tbl_auto_perceived_gender"
						}
					},
					"404": {
						"description": "NOT FOUND",
						"schema": {
							"$ref": "#/components/schemas/tbl_auto_perceived_gender"
						}
					}
				}
			},
			"put": {
				"tags": [
					"Basic data from auto perceived gender"
				],
				"parameters": [
					{
						"name": "ccn_auto_perceived_gender",
						"in": "path",
						"description": "todo id to update",
						"required": true,
						"type": "integer"
					}
				],
				"requestBody": {
					"description": "Update Existing todo On Database",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/tbl_auto_perceived_gender"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/components/schemas/tbl_auto_perceived_gender"
						}
					},
					"404": {
						"description": "NOT FOUND",
						"schema": {
							"$ref": "#/components/schemas/tbl_auto_perceived_gender"
						}
					}
				}
			},
			"delete": {
				"tags": [
					"Basic data from auto perceived gender"
				],
				"parameters": [
					{
						"name": "ccn_auto_perceived_gender",
						"in": "path",
						"description": "product id to delete",
						"required": true,
						"type": "integer"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/components/schemas/tbl_auto_perceived_gender"
						}
					},
					"404": {
						"description": "NOT FOUND",
						"schema": {
							"$ref": "#/components/schemas/tbl_auto_perceived_gender"
						}
					}
				}
			}
		},
...
		"/demographic_data": {
			"post": {
				"tags": [
					"Basic data from demographic data"
				],
				"description": "Create a demographic data data to human resource platform",
				"requestBody": {
					"description": "Post demographic data data to database",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/tbl_demographic_data"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "CREATED",
						"schema": {
							"$ref": "#/components/schemas/tbl_demographic_data"
						}
					},
					"404": {
						"description": "NOT FOUND",
						"schema": {
							"$ref": "#/components/schemas/tbl_demographic_data"
						}
					}
				}
			},
			"get": {
				"tags": [
					"Basic data from demographic data"
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/components/schemas/tbl_demographic_data"
						}
					},
					"404": {
						"description": "NOT FOUND",
						"schema": {
							"$ref": "#/components/schemas/tbl_demographic_data"
						}
					}
				}
			}
		},
		"/demographic_data/{ccn_demographic_data}": {
			"get": {
				"tags": [
					"Basic data from demographic data"
				],
				"parameters": [
					{
						"name": "ccn_demographic_data",
						"in": "path",
						"description": "todo id to update",
						"required": true,
						"type": "integer"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/components/schemas/tbl_demographic_data"
						}
					},
					"404": {
						"description": "NOT FOUND",
						"schema": {
							"$ref": "#/components/schemas/tbl_demographic_data"
						}
					}
				}
			},
			"put": {
				"tags": [
					"Basic data from demographic data"
				],
				"parameters": [
					{
						"name": "ccn_demographic_data",
						"in": "path",
						"description": "todo id to update",
						"required": true,
						"type": "integer"
					}
				],
				"requestBody": {
					"description": "Update Existing todo On Database",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/tbl_demographic_data"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/components/schemas/tbl_demographic_data"
						}
					},
					"404": {
						"description": "NOT FOUND",
						"schema": {
							"$ref": "#/components/schemas/tbl_demographic_data"
						}
					}
				}
			},
			"delete": {
				"tags": [
					"Basic data from demographic data"
				],
				"parameters": [
					{
						"name": "ccn_demographic_data",
						"in": "path",
						"description": "product id to delete",
						"required": true,
						"type": "integer"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/components/schemas/tbl_demographic_data"
						}
					},
					"404": {
						"description": "NOT FOUND",
						"schema": {
							"$ref": "#/components/schemas/tbl_demographic_data"
						}
					}
				}
			}
		},
...
	}
}